+++ WaveFrontParser/Model/Vertex.cs
@@ -7,9 +7,9 @@ namespace WaveFrontParser.Model
 {
     public class Vertex
     {
-        public double XAxis { get; set; }
-        public double YAxis { get; set; }
-        public double ZAxis { get; set; }
+        public float XAxis { get; set; }
+        public float YAxis { get; set; }
+        public float ZAxis { get; set; }
 
        
     }
